Create README.md
in this assignment (2.1), I had to make for my first assignment an initial with a gradient that every time the mouse passed a certain area the image would invert. In 2.2, I had to make a custom sprite that I would later move I made a guy with glasses on. In assignment 2.3, I made my little guy move one moved to the right indefinitely. I also, had 2 of the same sprite moving with my MouseX, MouseY. IN 2.4, I made a flag with two existing flags with Camaros and South Sudan. Thank you for reading.